decouple add alternative pt1
